restrictions,registration_page

+++ base/views.py
@@ -5,7 +5,7 @@ from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room,Topic
 from .forms import RoomForm
 
@@ -26,8 +26,14 @@ rooms =[
  """
 
 def loginPage(request):
+
+
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -46,7 +52,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or password is incorrect')
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html',context)
 
 
@@ -55,7 +61,21 @@ def logoutUser(request):
     return redirect('home')
     """ return redirect('logout') """
 
+def registerPage(request):
+  
+    form = UserCreationForm()
 
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occured during registration')
+    return render(request, 'base/login_register.html',{'form':form})
 
 def home(request):
 
@@ -78,8 +98,8 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-   
-    context = {'room':room}
+    room_messages = room.message_set.all().order_by('-created')
+    context = {'room':room,'room_messages':room_messages}
     return render(request, 'base/room.html',context,)
 
 @login_required(login_url='login')
@@ -117,6 +137,10 @@ def updateRoom(request,pk):
 
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed')
+        
     if request.method == 'POST':
         room.delete()
         return redirect('home')
